// src/model/CryptoProfite.ts
export interface CryptoProfite {
    CRY_PROFIT_ID?: number;
    TRAND_ID?: number;
    TRANS_DATE?: Date;
    ZAKAH_DATE?: Date;
    COIN_NAME?: string
    SPOT_PRICE?: number;
    NO_COINS?: number;
    TOT_STOCK_PRICE?: number;
    SALES_TAX?: number;
    TOT_MARKET_SALE_TAX?: number;
    TRANS_COST?: number;
    VALUE_BASED_CURR_PRICE?: number;
    PROFIT_VALUE?: number;
    PROFIT_PERCENT?: number;
    ZAKAH_VALUE?: number;
}